Users,Roles,Complains,Profile,User Logs, Holidays, Permissions and Suggestions Update By Muhammad Irfan.

+++ routes/holidays/holidays.php
@@ -4,19 +4,19 @@
 Route::group(['prefix' => 'Holidays', 'as' => 'holidays.', 'middleware' => ['auth:web']], function () {
 
 
-    Route::get('add', 'HolidayController@create')->name('add');
-    Route::post('store', 'HolidayController@store')->name('store');
+    Route::get('add', 'HolidayController@create')->name('add')->middleware('permission:add holiday,add');
+    Route::post('store', 'HolidayController@store')->name('store')->middleware('permission:add holiday,add');
 
-    Route::get('show', 'HolidayController@show')->name('show');
+    Route::get('show', 'HolidayController@show')->name('show')->middleware('permission:view holiday,view');
 
 //     Route::get('Edit', function () {
 //         return view('holidays.Edit');
 //     })->name('Edit');
 
-     Route::get('Edit/{id}', 'HolidayController@edit')->name('Edit');
-     Route::post('Edit/{id}', 'HolidayController@update')->name('Edit');
+     Route::get('Edit/{id}', 'HolidayController@edit')->name('Edit')->middleware('permission:edit holiday,Edit');
+     Route::post('Edit/{id}', 'HolidayController@update')->name('Edit')->middleware('permission:edit holiday,Edit');
 
-     Route::delete('delete/{id}', 'HolidayController@destroy')->name('delete');
+     Route::delete('delete/{id}', 'HolidayController@destroy')->name('delete')->middleware('permission:delete holiday,delete');;
